docker/app/backend/requirements.txt
flask
flask-cors
flask_compress
SQLAlchemy==2.0.17
python-dotenv
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.5